Enforce QtWayland's default DPI
Having a DPI but not a pixelDensity set causes some font sizes in the UI Toolkit to be multiplied twice: once by their internal scale factor based on GRID_UNIT_PX, and again by the logicalDpi value of the QScreen. On QtUbuntu, we were able to avoid this problem by implementing pixelDensity as a function of logicalDpi: https://github.com/ubports/qtubuntu/pull/5

Implementing pixelDensity is probably not a good option for QtWayland, so we will instead force its internal DPI to the default of 96 and allow our other scaling logic to do the job.